Ci/deploy all

+++ apps/chocolate-frontend/scripts/prebuild.sh
@@ -3,9 +3,13 @@
 CONTRACT_PATH=apps/chocolate-frontend/src/assets/contract/chocolate.json
 
 function fetch:contract(){
-  echo "CREATING AN EMPTY JSON AS CI STUB"
-  echo "{}" > $CONTRACT_PATH
-  touch $CONTRACT_PATH
+  if [ -s $CONTRACT_PATH ]; then
+    echo "File $CONTRACT_PATH is not empty"
+  else
+    echo "CREATING AN EMPTY JSON AS CI STUB"
+    echo "{}" > $CONTRACT_PATH
+    touch $CONTRACT_PATH
+  fi
 }
 
 "$@"
